# openapi.yaml
openapi: 3.0.0
info:
  title: Treza Platform API
  description: |
    Comprehensive API for managing secure enclaves, GitHub integrations, and platform resources.
    
    **New in v2.1.0:**
    - 🛡️ Cryptographic attestation documents and verification
    - 📋 Developer integration tools and API endpoints
    - 🏦 Use case templates for Financial, Healthcare, and Enterprise
    - 🔐 PCR measurements and security validation
    - 📊 Business impact and compliance tracking
    
    **Previous in v2.0.0:**
    - 🔄 Enclave lifecycle management (pause, resume, terminate)
    - 📜 Comprehensive logs access (application, ECS, Step Functions, Lambda, errors)
    - 🐳 Docker Hub integration (search images, get tags)
    - 📊 Enhanced status tracking with real-time updates
    - 🎯 Individual enclave management endpoints
  version: 2.1.0
  contact:
    name: Treza Support
    url: https://trezalabs.com
    email: support@trezalabs.com
servers:
  - url: https://app.trezalabs.com
    description: Production server
  - url: http://localhost:3000
    description: Development server
paths:
...
  /api/enclaves/{id}:
    get:
      summary: Get specific enclave
      description: Retrieve details for a specific enclave by ID
      tags:
        - Enclaves
      parameters:
        - name: id
          in: path
          required: true
          description: Enclave ID
          schema:
            type: string
            example: "enc_123456"
      responses:
        '200':
          description: Enclave details retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  enclave:
                    $ref: '#/components/schemas/Enclave'
        '404':
          description: Enclave not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    patch:
      summary: Manage enclave lifecycle
      description: Perform lifecycle actions on an enclave (pause, resume, terminate)
      tags:
        - Enclaves
      parameters:
        - name: id
          in: path
          required: true
          description: Enclave ID
          schema:
            type: string
            example: "enc_123456"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - action
                - walletAddress
              properties:
                action:
                  type: string
                  enum: [pause, resume, terminate]
                  example: "pause"
                  description: "Action to perform on the enclave"
                walletAddress:
                  type: string
                  example: "0x4B0897b0513fdc7C541B6d9D7E929C4e5364D2dB"
      responses:
        '200':
          description: Lifecycle action performed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  enclave:
                    $ref: '#/components/schemas/Enclave'
                  message:
                    type: string
                    example: "Enclave pause initiated successfully"
        '400':
          description: Invalid action or enclave state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized - wallet mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Enclave not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Delete specific enclave
      description: Delete a specific enclave by ID (only if in terminal state)
      tags:
        - Enclaves
      parameters:
        - name: id
          in: path
          required: true
          description: Enclave ID
          schema:
            type: string
            example: "enc_123456"
        - name: wallet
          in: query
          required: true
          description: Wallet address for authorization
          schema:
            type: string
            example: "0x4B0897b0513fdc7C541B6d9D7E929C4e5364D2dB"
      responses:
        '200':
          description: Enclave deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Enclave deleted successfully"
        '400':
          description: Enclave not in terminal state or missing parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Unauthorized - wallet mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Enclave not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
    Enclave:
      type: object
      properties:
        id:
          type: string
          example: "enc_123456"
        name:
          type: string
          example: "Trading Bot Enclave"
        description:
          type: string
          example: "Secure environment for automated trading strategies"
        status:
          type: string
          enum: [PENDING_DEPLOY, DEPLOYING, DEPLOYED, PAUSING, PAUSED, RESUMING, PENDING_DESTROY, DESTROYING, DESTROYED, FAILED]
          example: "PENDING_DEPLOY"
        region:
          type: string
          example: "us-east-1"
        walletAddress:
          type: string
          example: "0x4B0897b0513fdc7C541B6d9D7E929C4e5364D2dB"
        providerId:
          type: string
          example: "aws-nitro"
          description: "The ID of the provider used for this enclave"
        providerConfig:
          type: object
          example: {
            "dockerImage": "my-app:latest",
            "cpuCount": 2,
            "memoryMiB": 512
          }
          description: "Provider-specific configuration settings"
        createdAt:
          type: string
          format: date-time
          example: "2024-01-15T00:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2024-01-15T00:00:00Z"
        githubConnection:
          $ref: '#/components/schemas/GitHubConnection'
        error_message:
          type: string
          example: "Deployment failed: Instance type not compatible with Nitro Enclaves"
          description: "Error message if enclave failed (optional)"
...
    GitHubConnection:
      type: object
      properties:
        isConnected:
          type: boolean
          example: true
        username:
          type: string
          example: "example-user"
        selectedRepo:
          type: string
          example: "example-user/trading-bot"
        selectedBranch:
          type: string
          example: "main"
        accessToken:
          type: string
          example: "gho_xxxxxxxxxxxxxxxxxxxx"
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
        details:
          type: array
          items:
            type: string
          description: "Additional error details (optional)"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    
    WalletAuth:
      type: apiKey
      in: query
      name: wallet
      description: Wallet address for authentication
security:
  - BearerAuth: []
  - WalletAuth: []
...